#5 - Change requests, minor fixes

+++ src/main/java/ru/timerdar/FamilyFarmAPI/dto/Order.java
@@ -9,10 +9,10 @@ public class Order {
     private final String consumer_name;
     private final String  product_name;
     private final java.sql.Date start_data;
-    private final float amount;
+    private final double amount;
 
 
-    public Order(String consumer_name, String product_name, float amount) {
+    public Order(String consumer_name, String product_name, double amount) {
         this.consumer_name = consumer_name;
         this.product_name = product_name;
         Date d = new Date();
